requirements.txt
fastapi
uvicorn
google-generativeai
langchain
langchain-community
faiss-cpu
pypdf
requests
python-dotenv
sentence-transformers
